Turn some properties into categories, fix bug
There was a bug where `if IsUnitGroupOfGroupRing then` was written
instead of `if IsUnitGroupOfGroupRing(U) then`. However, after making
that fix, I realized that there are no methods installed for
`IsUnitGroupOfGroupRing`, ever. And indeed, thinking about it, I don't
believe it would make sense to do so, this filter indicates something
about how a group was constructed, and in general this can't be tested
post-hoc by a method. Thus, it seems sensible to turn change it from
a property to a category.

Perhaps for `IsNormalizedUnitGroupOfGroupRing` this would be possible,
but for now there is no method for it either, so I also change it to
a category.

Beware that this latter change will break unitlib; but a trivial change
there can remedy this.